Remove released keys from queue.
Upon release, the key is removed from queue, and keys later in the queue
are shifted to allow the slot to be reused. This avoids sending a second
release event for released keys when the hyper key is released. It also
fixes an issue where the queue can fill with released keys, preventing a
9th key from being queued.